+++ src/irida_wf_ga2xml/core.clj
@@ -58,10 +58,10 @@
   serialize into the expected IRIDA XML format for tool parameters."
   [step & {:keys [get-param-labels?
                   extra-tool-param-attrs?
-                  uuid-as-toolid?]
+                  use-tool-name-as-id?]
            :or   {get-param-labels?       true
                   extra-tool-param-attrs? false
-                  uuid-as-toolid?         false}}]
+                  use-tool-name-as-id?    false}}]
   (let [{:strs [tool_id id uuid]} step
         {:keys [name]} (tool-id->repo-info tool_id)
         params (tool-params-map step)
@@ -69,7 +69,7 @@
                       (msgs/get-tool-param-values step (keys params))
                       nil)
         xml-vec (vec (map (fn [[k v]]
-                            (let [base-attrs {:toolId        (if uuid-as-toolid? uuid tool_id)
+                            (let [base-attrs {:toolId        (if use-tool-name-as-id? tool_id uuid)
                                               :parameterName k}
                                   extra-attrs (if extra-tool-param-attrs?
                                                 (map (fn [x]
@@ -110,7 +110,7 @@
                           ^Boolean output-messages?
                           ^Boolean extra-tool-param-attrs?
                           ^Boolean remove-output-name-file-ext?
-                          ^Boolean uuid-as-toolid?]
+                          ^Boolean use-tool-name-as-id?]
                    :or   {single-sample?               true
                           wf-version                   "0.1.0"
                           analysis-type                "DEFAULT"
@@ -119,7 +119,7 @@
                           output-messages?             true
                           extra-tool-param-attrs?      false
                           remove-output-name-file-ext? false
-                          uuid-as-toolid?                    false}}]
+                          use-tool-name-as-id?         false}}]
   (let [ga-map (parse-ga path)
         {:strs [name annotation]} ga-map
         _ (info (str "Parsed Galaxy workflow file with name='" name "' and annotation/description='" annotation "'"))
@@ -133,7 +133,7 @@
                              (map #(tool-params-vec %
                                                     :get-param-labels? output-messages?
                                                     :extra-tool-param-attrs? extra-tool-param-attrs?
-                                                    :uuid-as-toolid? uuid-as-toolid?))
+                                                    :use-tool-name-as-id? use-tool-name-as-id?))
                              (filter not-empty))
         out {:name name
              :xml-vec