+++ CHANGELOG.md
@@ -3,6 +3,6 @@ Changes
 
 1.0.0 to 1.1.0
 --------------
-* [Feature]: Added option `-u`/`--uuid-as-toolid`to use uuid when generating tool IDs. This is needed when a workflow has multiple of the same tool.
+* [Feature]: Added option `-u`/`--use-tool-name-as-id` Use tool names when building tool_ids, will not work with workflows that have multiple of the same tool.
 * [Documentation]: Added information on how to get started with Leiningen